+++ window.h
@@ -29,7 +29,7 @@ typedef enum {
     NUMBER_OF_WINDOW_FUNCTIONS
 } window_function_t;
 
-extern double *get_window(window_function_t wfunc, int datalen);
+extern float *get_window(window_function_t wfunc, int datalen);
 extern void free_windows(void);
 extern const char *window_name(window_function_t wfunc);
 extern const char window_key(window_function_t wfunc);
